Made script to add buildings
